Multiple updates: - Added ARFilter and helper classes Polynomial and RationalPolynomial - Added debugData function to Extensions to allow for testing of ARFilter - Adjusted SampleFormat class: the field 'rate' was previously unused, is now used to store the amount of values contained in one packet for a given module. This to be able to process the output of the NexusSignal module in time domain, which contains more than one value per packet, in contrast to all previous modules. - Adjusted other sourceModules (GenerateSignal, KeyPressSignal, and PlaybackSignal) to fill in the correct value in the 'rate' field when registering their source streams. This was needed because of the adjustment of the SampleFormat class. - Adjusted NexusSignal to use the ARFilter when processing input in the time domain. Also included a debug mode that allows injecting test data, also when the Nexus can't find the Activa. Furthermore changed parameters, is now possible to define mappings from input to output channels, also for power mode. - Adjusted Data class to be able to correctly register and use adjusted SampleFormat, in order to process incoming sourcestreams that have more than one value per packet (needed for incoming data from NexusSignal module).
SVN rev 154

+++ UNP/src/Core/Helpers/SampleFormat.cs
@@ -8,14 +8,14 @@ namespace UNP.Core.Helpers {
     public class SampleFormat {
         
         private int channels = 0;                       //
-        private double rate = 0;                        // sample rate (in samples/second)
-        //private int[] types = null;                     // 
+        private int rate = 0;                        // amount of values contained in one packet
+        //private int[] types = null;                   // 
 
         public SampleFormat() {
         
         }
 
-        public SampleFormat(int channels, double rate) {
+        public SampleFormat(int channels, int rate) {
             this.channels = channels;
             this.rate = rate;
         }
@@ -24,7 +24,7 @@ namespace UNP.Core.Helpers {
             return channels;
         }
 
-        public double getRate() {
+        public int getRate() {
             return rate;
         }
